// types/milestone.ts
import { Timestamp } from "firebase/firestore";

export interface Milestone {
  id: string; // Unique ID for the milestone (e.g., generated by Firestore or nanoid)
  description: string;
  isCompleted: boolean;
  createdAt: Timestamp;
  updatedAt: Timestamp;
  // order?: number; // Optional: if you need to maintain a specific order
}